requirements.txt
Django==1.8.2
django-appconf==1.0.2
django-compressor==2.2
django-jsonfield==1.0.1
djangorestframework==3.1.2
jsonfield==2.0.2
rcssmin==1.0.6
rjsmin==1.0.12
